get_progressbar_data function moved to utils
