+++ trojsten/polls/templatetags/polls_parts.py
@@ -1,6 +1,6 @@
 from django import template
 
-from trojsten.utils.templatetags import progressbar
+from trojsten.utils import utils
 
 register = template.Library()
 
@@ -9,6 +9,6 @@ register = template.Library()
 def show_time(context, current_question):
     start = current_question.created_date
     end = current_question.deadline
-    data = progressbar.get_progressbar_data(start, end)
+    data = utils.get_progressbar_data(start, end)
     context.update({"current": current_question, **data})
     return context
